+++ tests/servlet_modules/application/servlets/loaded.jss.mjs
@@ -1,1 +1,1 @@
-export const value = "loadESM one";
+export const value = "ESM one";